+++ CHANGES.md
@@ -1,4 +1,5 @@
 - Add `ModelBase.primary_attrs` class property that returns a list of class attributes that are primary keys.
+- Use `ModelBase.primary_attrs` in `QueryModel.search()` so that it handles cases where primary keys have column names that are different than the class attribute name.
 
 ## v0.13.1 (2014-06-11)